Move Signal Connection to its own templated class
By not nesting Connection in SignalSource, it enables template
parameters to be deduced from constructor paramters, rather than forcing
developers to specify them when accessing the SignalSource namespace.
See unit test changes for an example of the impact of this.
